+++ lib/webspicy/client/rack_test_client.rb
@@ -5,13 +5,11 @@ module Webspicy
       Factory.new(app)
     end
 
-    def initialize(scope, app, domain)
+    def initialize(scope, app)
       super(scope)
       @api = Api.new(app)
-      @domain = domain
     end
     attr_reader :api
-    attr_reader :domain
 
     def call(test_case)
       service, resource = test_case.service, test_case.resource
@@ -22,7 +20,7 @@ module Webspicy
 
       # Instantiate the url and strip parameters
       url, params = resource.instantiate_url(params)
-      url = "http://#{domain}#{url}" if domain
+      url = scope.to_real_url(url){ url }
 
       # Invoke the service now
       api.public_send(service.method.to_s.downcase.to_sym, url, params, headers)
@@ -37,15 +35,9 @@ module Webspicy
         @app = app
       end
       attr_reader :app
-      attr_reader :domain
-
-      def on(domain)
-        @domain = domain
-        self
-      end
 
       def new(scope)
-        RackTestClient.new(scope, app, domain)
+        RackTestClient.new(scope, app)
       end
 
     end # class Factory